requirements.txt
absl-py==1.0.0
anyio==3.6.1
Appium-Python-Client==2.7.1
asgiref==3.5.2
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
convertapi==1.5.0
cycler==0.11.0
decorator==5.1.1
Django==4.0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
executing==1.1.1
fastapi==0.82.0
Flask==2.1.2
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h11==0.13.0
h5py==3.7.0
idna==3.3
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mysql-connector-python==8.0.23
mysqlclient==2.1.1
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.5.5.64
openpyxl==3.0.10
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
phonenumbers==8.12.56
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.31
protobuf==3.19.4
pure-eval==0.2.2
py4j==0.10.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==22.3.5
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pymongo==4.1.1
pyparsing==3.0.9
PySocks==1.7.1
pyspark==3.2.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
selenium==4.5.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stack-data==0.5.1
starlette==0.19.1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.1
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
uvicorn==0.18.3
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
wsproto==1.2.0
